# backend/api/openapi_0_2.yaml
openapi: 3.1.0
info:
  title: HomePlanner API v0.2
  description: >
    Спецификация HTTP API для релиза HomePlanner v0.2.
    Все временные поля передаются в локальном формате проекта без таймзоны (`YYYY-MM-DD HH:MM:SS`).
  version: 0.2.0
servers:
  - url: /api/v0.2
    description: Базовый префикс API v0.2
...
paths:
  /tasks:
    post:
      tags: [tasks]
      summary: Создание задачи
      description: Создаёт новую задачу с опциональными повторениями и напоминаниями.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Задача успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{taskId}:
    put:
      tags: [tasks]
      summary: Обновление задачи
      description: >
        Обновляет задачу. Конфликты разрешаются на сервере по времени обновления (`updated_at`).
        При синхронизации через `/tasks/sync-queue` операции с timestamp, который старше серверного `updated_at`, автоматически пропускаются.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Идентификатор задачи
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Задача успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизовано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Задача не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskBase:
      type: object
      required:
        - title
        - task_type
        - reminder_time
      properties:
        title:
          type: string
          description: Заголовок задачи
        description:
          type: string
          nullable: true
          description: Произвольное описание задачи
        task_type:
          type: string
          description: Тип задачи
          enum: [one_time, recurring, interval]
        recurrence_type:
          type: string
          nullable: true
          description: Тип повторяемости для расписаний
          enum: [daily, weekly, monthly, yearly, weekdays, weekends]
        recurrence_interval:
          type: integer
          minimum: 1
          nullable: true
          description: Интервал повторяемости (каждые N периодов)
        interval_days:
          type: integer
          minimum: 1
          nullable: true
          description: Длительность окна для интервальных задач (в днях)
        reminder_time:
          type: string
          format: date-time
          description: Дата и время напоминания в локальном формате проекта
          example: '2025-03-15T14:30:00'
        completed:
          type: boolean
          description: Выполнена ли задача в текущие сутки
          default: false
        active:
          type: boolean
          description: Активна ли задача (участвует в планировании и выдаче)
          default: true
        group_id:
          type: integer
          nullable: true
          description: Связанная группа задач (если есть)
        recurrence:
          $ref: '#/components/schemas/TaskRecurrencePayload'
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/TaskNotificationPayload'
    TaskCreateRequest:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        task_type:
          type: string
          enum: [one_time, recurring, interval]
        recurrence_type:
          type: string
          nullable: true
          enum: [daily, weekly, monthly, yearly, weekdays, weekends]
        recurrence_interval:
          type: integer
          minimum: 1
          nullable: true
        interval_days:
          type: integer
          minimum: 1
          nullable: true
        reminder_time:
          type: string
          format: date-time
        completed:
          type: boolean
        active:
          type: boolean
        group_id:
          type: integer
          nullable: true
        recurrence:
          $ref: '#/components/schemas/TaskRecurrencePayload'
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/TaskNotificationPayload'
        soft_lock_owner:
          type: string
          nullable: true
        soft_lock_expires_at:
          type: string
          format: date-time
          nullable: true
    TaskResponse:
      type: object
      required:
        - id
        - title
        - task_type
        - reminder_time
        - completed
        - active
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        task_type:
          type: string
          enum: [one_time, recurring, interval]
        recurrence_type:
          type: string
          nullable: true
          enum: [daily, weekly, monthly, yearly, weekdays, weekends]
        recurrence_interval:
          type: integer
          nullable: true
        interval_days:
          type: integer
          nullable: true
        reminder_time:
          type: string
          format: date-time
        completed:
          type: boolean
        active:
          type: boolean
        group_id:
          type: integer
          nullable: true
        last_shown_at:
          type: string
          format: date-time
          nullable: true
        readable_config:
          type: string
          nullable: true
        recurrence:
          $ref: '#/components/schemas/TaskRecurrenceResponse'
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/TaskNotificationResponse'
        soft_lock_owner:
          type: string
          nullable: true
        soft_lock_expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskRecurrencePayload:
      type: object
      properties:
        rrule:
          type: string
          description: Правило повторяемости в формате RRULE
          example: 'FREQ=WEEKLY;BYDAY=MO,WE,FR'
        end_at:
          type: string
          format: date-time
          nullable: true
          description: Дата завершения повторов (локальный формат)
    TaskRecurrenceResponse:
      allOf:
        - $ref: '#/components/schemas/TaskRecurrencePayload'
      type: object
      required:
        - id
        - rrule
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskNotificationPayload:
      type: object
      properties:
        notification_type:
          type: string
          description: Тип уведомления (например, reminder)
          example: reminder
        channel:
          type: string
          description: Канал доставки (push, local и т.д.)
          example: local
        offset_minutes:
          type: integer
          description: Смещение в минутах относительно времени напоминания
          example: 15
    TaskNotificationResponse:
      allOf:
        - $ref: '#/components/schemas/TaskNotificationPayload'
      type: object
      required:
        - id
        - notification_type
        - channel
        - offset_minutes
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string